requirements.txt
Django==2.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
Pillow==7.2.0
pytz==2020.1
